# backend/src/entities/message/types.ts
export type MessageRole = 'user' | 'assistant'

export interface Message {
  id: string
  conversationId: string
  role: MessageRole
  content: string
  reasoning: string | null
  createdAt: string
}